requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-migrate>=4.0.5
flask-wtf
python-dotenv
pymysql>=1.1.0
wtforms[email]
requests
flask-login
Flask-Babel
flask-session
wtforms-sqlalchemy
flask-limiter
redis>=5.0.0  # Required for rate limiting in production (Opcja 1)
isbnlib
Pillow
APScheduler
python-json-logger
flask-caching